Create Printer MIB supply query
This script queries printer MIBs for information on toner levels, and prints them in the shell.
the application snmpwalk is a prerequisite for this script to run. ensure the exe is available to PS.
Note to decent responsible admins, consult snmpwalk parameters for SNMPv3.
Vendors do NOT have to use these OIDs so your milage may vary, but DEVs are lazy so I've had good luck.